Refactored get_summary EP in admin_analytics
